Rename `foo/mod.rs` files too `foo.rs`
